Address review comments about doc and rename category to auxiliary
