+++ test/unit/CMakeLists.txt
@@ -25,6 +25,6 @@ add_subdirectory(matrix)
 add_subdirectory(memory)
 
 # Operations
+add_subdirectory(auxiliary)
 add_subdirectory(factorization)
 add_subdirectory(solver)
-add_subdirectory(utility)
